devtool: ide: add install dependency before gdbserver task
On VSCode, when running the debugger through launch.json, the program is
recompiled, gdbserver is started, and connected to. However, the task to
install the program on the target is not run, so the program is not
updated on the target.

Adding a dependency to the `install && deploy-target` task fixes this.