devtool: ide: remove oe-local-files from vscode excludes
With the cmake-example recipe, the oe-local-files directory is
actually used when browsing or debugging the source code. The reason
is that VSCode automatically dereferences the symlinks present in the
workspace. Since the user will often be opening the symlinks and the
hard files, we should let him freely navigate them.