devtool: ide_sdk: Use bitbake python3 for generated scripts
The generated scripts use the sys.path configuration used inside
bitbake. It can be a different python version than the one used on the
host. However, the generated script is intended to run in a different
host IDE environment.

For instance, when running the generated script
deploy_target_cmake-example-core2-64 from an eSDK generated on another
machine, I got the following exception:
    AssertionError: SRE module mismatch

We need to match the sys.executable to the sys.path.