devtool: code: Provide a generic C++ configuration
By default, the cpptools VSCode extension will use the host's headers
and flags for linting. This results in a lot of include errors and
misleading definitions. Even though this generic configuration doesn't
include all the depenendencies, it is a proper fallback for recipe
classes we do not accurately cover, with at least the right sysroot.

Additionally, ide-sdk automatically detects and provides a launch.json
configuration for autotools recipes so we should recommend the C++
extensions for them.

If the recipe is of another class (like Python), then the configuration
will be generated but probably useless to the user. We'd need a mecanism
to know if the recipe contains C/C++ but I'm not aware of any
class-agnostic way to do it.