image.bbclass/rootfs: set and unpack package-database
set the package-database of a "lower image" to unpack and build upon when
installing packages for the current image. This way a lean image will be
created, which only holds the packages that are not already present in the lower
image, that then could be used with overlayfs or systemd-sysext to extend the
"lower image" on demand; for development purposes on an RO lower image for
example.

(From OE-Core rev: 1df040d1688534ce99055baeb0a988d4e5398452)

Signed-off-by: Johannes Schneider <johannes.schneider@leica-geosystems.com>
Signed-off-by: Alexandre Belloni <alexandre.belloni@bootlin.com>